Warp now uses gal_data_t and all its fruits
The Warp program now uses all the new features offered by `gal_data_t' in
this branch. Several major changes were made to make it more user friendly
as described below:

 - The old `--nowcscorrection' option has been given a more clear name of
   `--keepwcs'. With this option, Warp will not apply the warp the input's
   WCS structure.

 - The old `--maxblankfrac' option has been changed to
   `--coveredfrac'. Until now, Warp would only look for the fraction of
   input blank/NaN pixel area over the output pixel. But this would be
   useless on the edges of the image. So the new `--coveredfrac' option
   takes the acceptable fraction of output pixel area that must be covered
   by input pixels in order to give that output pixel a value. When given a
   sufficiently high value (1.0 is the largest possible), it will allow the
   user to not be bothered by bad edge values which are caused because of
   not having any data outside the image.

 - The old `--nofitscorrect' option has been changed to `--centeroncorner'
   to be more clear. The new option is now more general than
   before. Before, it would only work on 2D matrixes and modular warpings,
   but now, it is general and happens after the final input Matrix is
   ready, so it even applies to 3D raw input matrix. The previous case was
   confusing both for the users and developers.

 - The old `--zerofornoinput' option has been removed. If the user doesn't
   want blank pixels for regions were there was no data, they can use
   Arithmetic.

 - The old `--doubletype' option has been removed, since the common option
   `--type' now comes in from the infra-structure which is much more
   flexible and allows the user to set any type, not just double.

 - A new `func' element was added to `argp_option' that will be used by the
   option parsers to immediately parse the value and store the value in a
   lower-level way that can be used by the programs more efficiently. This
   was mainly done because of the modular warpings of Warp: they have
   differnt option names, but because order matters in them, they must all
   be kept in one lower-level linked list. After its implementation, it is
   now also used for the common options that take strings for values that
   are internally stored as integer constants: the common options `--type',
   `--tableformat', and `--searchin'. When writing the options and their
   values, it will also use the same function to retrieve the static
   string. Unfortunately for the time being, it doesn't print the values of
   the modular warpings of Warp, but later we can implement it, maybe the
   matrix option alone (which contains all the warping information) would
   be enough.

This completes task #14365 and task #14368.

+++ lib/fits.c
@@ -219,7 +219,7 @@ gal_fits_bitpix_to_type(int bitpix)
 
 
 int
-gal_fits_type_to_bitpix(int type)
+gal_fits_type_to_bitpix(uint8_t type)
 {
   switch(type)
     {
@@ -258,7 +258,7 @@ gal_fits_type_to_bitpix(int type)
    of the column. So this function will do the conversion based on the
    CFITSIO manual.*/
 char
-gal_fits_type_to_bin_tform(int type)
+gal_fits_type_to_bin_tform(uint8_t type)
 {
   switch(type)
     {
@@ -300,7 +300,7 @@ gal_fits_type_to_bin_tform(int type)
 
 
 int
-gal_fits_type_to_datatype(int type)
+gal_fits_type_to_datatype(uint8_t type)
 {
   int w=0;
 
@@ -790,7 +790,7 @@ gal_fits_key_read(char *filename, char *hdu, gal_data_t *keysll,
    it is important to know before hand if they were allocated or
    not. If not, they don't need to be freed. */
 void
-gal_fits_key_add_to_ll(struct gal_fits_key_ll **list, int type,
+gal_fits_key_add_to_ll(struct gal_fits_key_ll **list, uint8_t type,
                        char *keyname, int kfree, void *value, int vfree,
                        char *comment, int cfree, char *unit)
 {
@@ -820,7 +820,7 @@ gal_fits_key_add_to_ll(struct gal_fits_key_ll **list, int type,
 
 
 void
-gal_fits_key_add_to_ll_end(struct gal_fits_key_ll **list, int type,
+gal_fits_key_add_to_ll_end(struct gal_fits_key_ll **list, uint8_t type,
                            char *keyname, int kfree, void *value, int vfree,
                            char *comment, int cfree, char *unit)
 {
@@ -1237,7 +1237,7 @@ gal_fits_img_read(char *filename, char *hdu, size_t minmapsize)
    this input to be a special type. For such cases, this function can be
    used to convert the input file to the desired type. */
 gal_data_t *
-gal_fits_img_read_to_type(char *inputname, char *hdu, int type,
+gal_fits_img_read_to_type(char *inputname, char *hdu, uint8_t type,
                           size_t minmapsize)
 {
   gal_data_t *in, *converted;
@@ -1340,6 +1340,7 @@ gal_fits_img_write_to_ptr(gal_data_t *data, char *filename)
     fits_create_file(&fptr, filename, &status);
 
   /* Create the FITS file. */
+
   fits_create_img(fptr, gal_fits_type_to_bitpix(data->type),
                   data->ndim, naxes, &status);
   gal_fits_io_error(status, NULL);
